Kaplan-Meier plot for events

Usage

kmplot(
  nif,
  analyte,
  dose = NULL,
  group = NULL,
  convert_tafd_h_to_d = TRUE,
  title = NULL,
  y_label = NULL,
  show_censoring = TRUE,
  show_risk_table = TRUE,
  show_ci = TRUE,
  silent = NULL,
  ...
)

Arguments

nif

A nif object.

analyte

The analyte.

dose

The dose, defaults to all doses, if NULL.

group

Grouping variable.

convert_tafd_h_to_d

Convert the TAFD field from hours to days, defaults to TRUE.

title

The plot title, defaults to none if NULL.

y_label

The y axis label, defaults to the analyte, if NULL.

show_censoring

Add censor mark, as logical.

show_risk_table

Add risk table, as logical.

show_ci

Add confidence interval, as logical.

silent

Suppress messages, as logical. Defaults to nif_option setting if NULL.

...

Arguments passed on to survminer::ggsurvplot

conf.int

logical value. If TRUE, plots confidence interval.

pval

logical value, a numeric or a string. If logical and TRUE, the p-value is added on the plot. If numeric, than the computet p-value is substituted with the one passed with this parameter. If character, then the customized string appears on the plot. See examples - Example 3.

surv.median.line

character vector for drawing a horizontal/vertical line at median survival. Allowed values include one of c("none", "hv", "h", "v"). v: vertical, h:horizontal.

risk.table

Allowed values include:

TRUE or FALSE specifying whether to show or not the risk table. Default is FALSE.

"absolute" or "percentage". Shows the absolute number and the percentage of subjects at risk by time, respectively.

"abs_pct" to show both absolute number and percentage.

"nrisk_cumcensor" and "nrisk_cumevents". Show the number at risk and, the cumulative number of censoring and events, respectively.

Value

A ggplot object.